% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bootstrap_dist}
\alias{bootstrap_dist}
\title{runs bootstrapping of ranks}
\usage{
bootstrap_dist(df, quo_score_col, quo_group_col, nits = 10, control = "A")
}
\arguments{
\item{df}{input data frame}

\item{quo_score_col}{quoted column name containing the HR scores}

\item{quo_group_col}{quoted group name containing the group column}

\item{nits}{number of bootstrap iterations to do}

\item{control}{character naming the control group that will be removed prior
to bootstrapping}
}
\description{
\code{boostrap_dist} runs multiple bootstrap resamplings using
\code{\link{bstrap_sample}} to generate a data frame of bootstrap rank means
of per group scores
}
\keyword{internal}
