% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bstrap_sample}
\alias{bstrap_sample}
\title{resamples a rank column in a dataframe based on groups}
\usage{
bstrap_sample(iteration, quo_score_col, quo_group_col, df, control = "A")
}
\arguments{
\item{iteration}{integer giving the current bootstrap iteration from the
calling function}

\item{quo_score_col}{quoted column name containing the HR scores}

\item{quo_group_col}{quoted group name containing the group name of the
observation}

\item{df}{input data frame}

\item{control}{character naming the control group that will be removed prior
to bootstrapping}
}
\description{
\code{bstrap_sample} performs a single iteration of bootstrapping of a rank
column in a data frame. The rows matching \code{control} in the quoted
\code{quo_group_col} column are removed. The remaining rows are grouped by
\code{quo_group_col} and in each group the rank column is resampled with
replacement and the mean rank of the resampling for each group is returned
}
\keyword{internal}
