% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-panels.R
\name{build_bootstrap_panel}
\alias{build_bootstrap_panel}
\title{Build the bootstrap distribution panel}
\usage{
build_bootstrap_panel(data_view, config)
}
\arguments{
\item{data_view}{A besthr_data_view object}

\item{config}{A besthr_plot_config object}
}
\value{
A ggplot object
}
\description{
Creates a ggplot showing ridge density plots of bootstrap distributions with
confidence interval shading.
}
\examples{
d <- make_data()
hr <- estimate(d, score, group)
dv <- besthr_data_view(hr)
cfg <- besthr_plot_config()

build_bootstrap_panel(dv, cfg)

}
