% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-panels.R
\name{build_ranked_panel}
\alias{build_ranked_panel}
\title{Build the ranked observation panel}
\usage{
build_ranked_panel(data_view, config)
}
\arguments{
\item{data_view}{A besthr_data_view object}

\item{config}{A besthr_plot_config object}
}
\value{
A ggplot object
}
\description{
Internal function to build the panel showing ranked observations.
}
\keyword{internal}
