% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-panels.R
\name{build_tech_rep_panel}
\alias{build_tech_rep_panel}
\title{Build the technical replicate panel}
\usage{
build_tech_rep_panel(data_view, config)
}
\arguments{
\item{data_view}{A besthr_data_view object}

\item{config}{A besthr_plot_config object}
}
\value{
A ggplot object
}
\description{
Internal function to build the panel showing raw scores with tech reps.
}
\keyword{internal}
