% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-panels.R
\name{compose_besthr_panels}
\alias{compose_besthr_panels}
\title{Compose besthr panels}
\usage{
compose_besthr_panels(panels, config)
}
\arguments{
\item{panels}{A list of ggplot objects to compose}

\item{config}{A besthr_plot_config object}
}
\value{
A patchwork object
}
\description{
Combines observation and bootstrap panels using patchwork with proper
alignment and shared legends.
}
\examples{
d <- make_data()
hr <- estimate(d, score, group)
dv <- besthr_data_view(hr)
cfg <- besthr_plot_config()

p1 <- build_observation_panel(dv, cfg)
p2 <- build_bootstrap_panel(dv, cfg)
compose_besthr_panels(list(p1, p2), cfg)

}
