% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{conf_intervals}
\alias{conf_intervals}
\title{gets confidence interval limits for means of bootstrapped ranks}
\usage{
conf_intervals(df, quo_group_col, low = 0.05, high = 0.95)
}
\arguments{
\item{df}{input data frame}

\item{quo_group_col}{quoted group name containing the column to group on}

\item{low}{the low probability value of the quantile}

\item{high}{the high probability value of the quantile}
}
\description{
\code{conf_intervals} calculates a low and a high quantile for the
\code{mean} column of each group of a dataframe
}
\keyword{internal}
