% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-hrest.R
\name{dot_plot}
\alias{dot_plot}
\title{dot plot of ranked data without technical replicates}
\usage{
dot_plot(hrest, group_col, theme_style = "modern", color_palette = "okabe_ito")
}
\arguments{
\item{hrest}{the hrest object from \code{estimate}}

\item{group_col}{quoted group column name}

\item{theme_style}{character specifying the theme style}

\item{color_palette}{character specifying the color palette}
}
\description{
\code{dot_plot} returns a ggplot object of ranked data with group on the
x-axis and rank on the y-axis. Point size indicates the number of
observations seen at that point. A per group horizontal line shows the group
ranked mean
}
\keyword{.data}
\keyword{importFrom}
\keyword{internal}
\keyword{rlang}
