% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{factorise_cols}
\alias{factorise_cols}
\title{Convert named columns to factors}
\usage{
factorise_cols(df, col_list)
}
\arguments{
\item{df}{data frame}

\item{col_list}{vector of quoted column names}
}
\description{
\code{factorise_cols} returns a copy of the passed data frame in which all
explicitly named columns are converted to factors with \code{as.factor}
All columns with only NA values are ommitted.
}
\keyword{internal}
