% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{format_table}
\alias{format_table}
\title{Format table as string}
\usage{
format_table(tbl, format)
}
\arguments{
\item{tbl}{A tibble to format}

\item{format}{The output format}
}
\value{
A character string
}
\description{
Internal function to convert tibble to markdown, html, or latex string.
}
\keyword{internal}
