% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{group_means}
\alias{group_means}
\title{Get mean rank of groups in a data frame}
\usage{
group_means(df, quo_group_col)
}
\arguments{
\item{df}{input data frame}

\item{quo_group_col}{quoted group name containing the column to group on}
}
\description{
\code{group_means} groups the provided dataframe by a column and returns
a summary dataframe with a mean column containing the mean of the group's
rank column
}
\keyword{internal}
