% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{group_ns}
\alias{group_ns}
\title{Get number of observations in a group in a data frame}
\usage{
group_ns(df, quo_group_col)
}
\arguments{
\item{df}{input data frame}

\item{quo_group_col}{quoted group name containing the column to group on}
}
\description{
\code{group_ns} groups the provided dataframe by a column and returns
a summary dataframe with an column \code{n} containg the number of
observations in a group
}
\keyword{internal}
