% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-layers.R
\name{layer_bootstrap_density}
\alias{layer_bootstrap_density}
\title{Create bootstrap density layer}
\usage{
layer_bootstrap_density(data_view, config)
}
\arguments{
\item{data_view}{A besthr_data_view object}

\item{config}{A besthr_plot_config object}
}
\value{
A list of ggplot2 layers
}
\description{
Creates a ggplot2 layer showing ridge density plots of bootstrap distributions
with confidence interval shading.
}
\examples{
\donttest{
d <- make_data()
hr <- estimate(d, score, group)
dv <- besthr_data_view(hr)
cfg <- besthr_plot_config()
# layer_bootstrap_density returns layers to add to a ggplot
}

}
