% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-layers.R
\name{layer_group_means}
\alias{layer_group_means}
\title{Create group mean lines layer}
\usage{
layer_group_means(data_view, config)
}
\arguments{
\item{data_view}{A besthr_data_view object}

\item{config}{A besthr_plot_config object}
}
\value{
A ggplot2 layer
}
\description{
Creates a ggplot2 layer showing horizontal lines at group mean ranks.
}
\examples{
\donttest{
d <- make_data()
hr <- estimate(d, score, group)
dv <- besthr_data_view(hr)
cfg <- besthr_plot_config()
# layer_group_means returns a ggplot layer
}

}
