% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-layers.R
\name{layer_ranked_dots}
\alias{layer_ranked_dots}
\title{Create ranked dots layer}
\usage{
layer_ranked_dots(data_view, config)
}
\arguments{
\item{data_view}{A besthr_data_view object}

\item{config}{A besthr_plot_config object}
}
\value{
A list of ggplot2 layers
}
\description{
Creates a ggplot2 layer showing ranked observations as points, where point
size indicates the count of observations at each rank/group combination.
}
\examples{
\donttest{
d <- make_data()
hr <- estimate(d, score, group)
dv <- besthr_data_view(hr)
cfg <- besthr_plot_config()
# layer_ranked_dots returns layers to add to a ggplot
}

}
