% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-layers.R
\name{layer_tech_rep_dots}
\alias{layer_tech_rep_dots}
\title{Create technical replicate dots layer}
\usage{
layer_tech_rep_dots(data_view, config)
}
\arguments{
\item{data_view}{A besthr_data_view object}

\item{config}{A besthr_plot_config object}
}
\value{
A list of ggplot2 layers
}
\description{
Creates a ggplot2 layer showing raw score observations with technical
replicates displayed separately. Points are sized by observation count.
}
