% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{list_besthr_styles}
\alias{list_besthr_styles}
\title{List available style presets}
\usage{
list_besthr_styles()
}
\value{
A character vector of style names (invisibly)
}
\description{
Shows all available preset styles that can be used with \code{besthr_style()}.
}
\examples{
list_besthr_styles()

}
