% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_data}
\alias{make_data}
\title{return a sample data set of random values for two groups}
\usage{
make_data()
}
\value{
tibble of random values for two groups
}
\description{
return a sample data set of random values for two groups
}
\examples{

 d1 <- make_data()

}
