% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_data2}
\alias{make_data2}
\title{return a sample data set of random values for two groups with three technical reps per group}
\usage{
make_data2()
}
\value{
tibble of random values for two groups with three technical reps per group
}
\description{
return a sample data set of random values for two groups with three technical reps per group
}
\examples{

 d2 <- make_data2()

}
