% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_data3}
\alias{make_data3}
\title{return a sample data set of random values for three groups with three technical reps per group}
\usage{
make_data3()
}
\value{
tibble of random values for three groups with three technical reps per group
}
\description{
@examples
}
\details{
d3 <- make_data3()
}
