% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-layers.R
\name{print.besthr_data_view}
\alias{print.besthr_data_view}
\title{Print method for besthr_data_view}
\usage{
\method{print}{besthr_data_view}(x, ...)
}
\arguments{
\item{x}{A besthr_data_view object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print method for besthr_data_view
}
