% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-config.R
\name{print.besthr_plot_config}
\alias{print.besthr_plot_config}
\title{Print method for besthr_plot_config}
\usage{
\method{print}{besthr_plot_config}(x, ...)
}
\arguments{
\item{x}{A besthr_plot_config object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print method for besthr_plot_config
}
