% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.hrest}
\alias{print.hrest}
\title{print a summary of the hrest object}
\usage{
\method{print}{hrest}(x, ...)
}
\arguments{
\item{x}{hrest object}

\item{...}{other parameters}
}
\value{
null
}
\description{
print a summary of the hrest object
}
\examples{

 d1 <- make_data()
 hr_est <- estimate(d1, score, group)
 print(hr_est)

}
