% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{scale_color_besthr}
\alias{scale_color_besthr}
\alias{scale_colour_besthr}
\title{Discrete color scale for besthr}
\usage{
scale_color_besthr(palette = "default", ...)

scale_colour_besthr(palette = "default", ...)
}
\arguments{
\item{palette}{Character string specifying the palette (see \code{\link{besthr_palette}})}

\item{...}{Additional arguments passed to \code{\link[ggplot2]{discrete_scale}}}
}
\value{
A ggplot2 discrete color scale
}
\description{
A discrete color scale using besthr palettes.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, wt, color = factor(cyl))) +
  geom_point() +
  scale_color_besthr("okabe_ito")

}
