% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{scale_fill_besthr}
\alias{scale_fill_besthr}
\title{Discrete fill scale for besthr}
\usage{
scale_fill_besthr(palette = "default", ...)
}
\arguments{
\item{palette}{Character string specifying the palette (see \code{\link{besthr_palette}})}

\item{...}{Additional arguments passed to \code{\link[ggplot2]{discrete_scale}}}
}
\value{
A ggplot2 discrete fill scale
}
\description{
A discrete fill scale using besthr palettes.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(factor(cyl), fill = factor(cyl))) +
  geom_bar() +
  scale_fill_besthr("okabe_ito")

}
