% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-hrest.R
\name{tech_rep_dot_plot}
\alias{tech_rep_dot_plot}
\title{dot plot of score data with technical replicates}
\usage{
tech_rep_dot_plot(
  hrest,
  score_col,
  group_col,
  tech_rep_col,
  theme_style = "modern",
  color_palette = "okabe_ito"
)
}
\arguments{
\item{hrest}{the hrest object from \code{estimate}}

\item{score_col}{quoted score column name}

\item{group_col}{quoted group column name}

\item{tech_rep_col}{quoted tech replicate column name}

\item{theme_style}{character specifying the theme style}

\item{color_palette}{character specifying the color palette}
}
\description{
\code{tech_rep_dot_plot} returns a ggplot object of score data with group on
technical replicate on the x-axis, score on the y-axis with point size
representing the number of observations at that point. Facets represent
individual groups
}
\keyword{.data}
\keyword{ImportFrom}
\keyword{internal}
\keyword{rlang}
