% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-config.R
\name{update_config}
\alias{update_config}
\title{Update a besthr plot configuration}
\usage{
update_config(config, ...)
}
\arguments{
\item{config}{An existing besthr_plot_config object}

\item{...}{Named arguments to update}
}
\value{
A new besthr_plot_config object
}
\description{
Creates a new configuration by updating specific fields of an existing one.
}
\examples{
cfg <- besthr_plot_config()
cfg2 <- update_config(cfg, theme_style = "modern", panel_widths = c(2, 1))

}
