% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_rank}
\alias{add_rank}
\title{calculate ranks of a score column and adds result to a dataframe}
\usage{
add_rank(df, quo_score_col)
}
\arguments{
\item{df}{data frame}

\item{quo_score_col}{quoted score column name}
}
\description{
\code{add_rank} returns a copy of the passed data frame containing a new
 column called \code{rank} which has the ranks of the named score column
}
\keyword{internal}
